Rename file from model.py
